requirements.txt

django-crispy-forms==1.9.2
Pillow==7.2.0